refactor(files): split FileTreeOverlay under the 500-line component cap
Extract FileTreeRow into its own module and the drag-resize logic into
use-tree-panel-resize.ts. No behavior change; fixes the repo-shape
max-lines check (541 > 500).
